requirements.txt
Flask
Werkzeug
azure-ai-formrecognizer
azure-core
python-dotenv